Move UI related code into ui package
